+++ app/src/main/java/com/example/paymenttracker/LendTransaction.java
@@ -1,11 +1,12 @@
 package com.example.paymenttracker;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class LendTransaction extends Transaction {
     private boolean isOutstanding;
 
-    LendTransaction(String name, String description, Date date, float amount, int category, int type) {
+    LendTransaction(String name, String description, Date date, BigDecimal amount, int category, int type) {
         super(name, description, date, amount, category, type);
         isOutstanding = true;
     }
